Adicionado o módulo de atividade conforme solicitado pelo cliente feita as modificações necessárias para funcionar corretamente

+++ framework/controle/Crud.php
@@ -10,10 +10,7 @@ class Crud extends Auxiliar {
     protected $nomeProduto;
 
     public function acaoFK() {
-//        $PRODUTOS = array();
-//        foreach ($_POST["PRO_CODIGO"] as $key => $value) {
-//            $PRODUTOS["CODIGO"][$key] = $value;
-//        }
+        $seq =1;
         foreach ($this->multi as $tabela => $chave) {
             $DADOS = array();
             $DADOS[$this->chave] = fCoalesce($this->chaveValor, $this->getValorChave());
@@ -35,21 +32,27 @@ class Crud extends Auxiliar {
                     }
                     $DADOS[$coluna] = $kkkkk;
                 }
-                for ($i = 0; $i < count($DADOS["PRO_CODIGO"]); $i++) {
-                    $this->setSql("SELECT * FROM PRODUTO_ITEM WHERE PRO_CODIGO = {$DADOS["PRO_CODIGO"]} ");
-                    foreach ($this->consulta() as $value) {
-                        $quantidade_peca = floatval($_POST["PRO_QUANTIDADE"][$indice]) * floatval($value["ITE_QUANTIDADE"]);
-                        $this->setSql("SELECT * FROM TAREFA_ITEM WHERE TAR_CODIGO = {$DADOS["TAR_CODIGO"]} AND ITE_CODIGO = {$value["ITE_CODIGO"]} ");
-                        foreach ($this->consulta() as $ITENS) {
-                            if ($ITENS != null) {
-                                $quantidade_peca = $quantidade_peca + $ITENS["ITE_QUANTIDADE"];
-                                $this->setSql("DELETE FROM TAREFA_ITEM WHERE TAR_CODIGO = {$DADOS["TAR_CODIGO"]} AND ITE_CODIGO = {$value["ITE_CODIGO"]} ");
-                                $this->consulta();
+                if ($this->modulo == "tarefa") {
+                    for ($i = 0; $i < count($DADOS["PRO_CODIGO"]); $i++) {
+                        $this->setSql("SELECT * FROM PRODUTO_ITEM WHERE PRO_CODIGO = {$DADOS["PRO_CODIGO"]} ");
+                        foreach ($this->consulta() as $value) {
+                            $quantidade_peca = floatval($_POST["PRO_QUANTIDADE"][$indice]) * floatval($value["ITE_QUANTIDADE"]);
+                            $this->setSql("SELECT * FROM TAREFA_ITEM WHERE TAR_CODIGO = {$DADOS["TAR_CODIGO"]} AND ITE_CODIGO = {$value["ITE_CODIGO"]} ");
+                            foreach ($this->consulta() as $ITENS) {
+                                if ($ITENS != null) {
+                                    $quantidade_peca = $quantidade_peca + $ITENS["ITE_QUANTIDADE"];
+                                    $this->setSql("DELETE FROM TAREFA_ITEM WHERE TAR_CODIGO = {$DADOS["TAR_CODIGO"]} AND ITE_CODIGO = {$value["ITE_CODIGO"]} ");
+                                    $this->consulta();
+                                }
                             }
+                            $this->setSql("INSERT INTO TAREFA_ITEM (TAR_CODIGO, ITE_CODIGO, ITE_QUANTIDADE) VALUES ({$DADOS["TAR_CODIGO"]}, {$value["ITE_CODIGO"]}, $quantidade_peca)");
+                            $this->consulta();
                         }
-                        $this->setSql("INSERT INTO TAREFA_ITEM (TAR_CODIGO, ITE_CODIGO, ITE_QUANTIDADE) VALUES ({$DADOS["TAR_CODIGO"]}, {$value["ITE_CODIGO"]}, $quantidade_peca)");
-                        $this->consulta();
                     }
+                }
+                if ($tabela == "PRODUTO_ATIVIDADE") {
+                    $DADOS["PRD_SEQUENCIA"] = $seq;
+                    $seq++;
                 }
                 $this->inserir($tabela, $DADOS);
             }
